Move cobrowse logic to its own file
