+++ src/client/components/ListPagePaginated/ListPagePaginated.tsx
@@ -4,7 +4,7 @@ import { Paragraph } from '@amsterdam/design-system-react';
 import { useParams } from 'react-router';
 
 import { LinkProps, ZaakDetail } from '../../../universal/types/App.types';
-import { getRedactedClass } from '../../helpers/utils';
+import { getRedactedClass } from '../../helpers/cobrowse';
 import { usePageTypeSetting } from '../../hooks/useThemaMenuItems';
 import ErrorAlert from '../Alert/Alert';
 import LoadingContent from '../LoadingContent/LoadingContent';
