+++ src/client/components/Thema/ThemaDetailPagina.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 
 import { LinkProps, ZaakDetail } from '../../../universal/types/App.types';
-import { getRedactedClass } from '../../helpers/utils';
+import { getRedactedClass } from '../../helpers/cobrowse';
 import ErrorAlert from '../Alert/Alert';
 import LoadingContent, { BarConfig } from '../LoadingContent/LoadingContent';
 import { DetailPageV2, PageContentCell, PageContentV2 } from '../Page/Page';
