+++ src/client/helpers/utils.ts
@@ -1,9 +1,5 @@
 import { KeyboardEvent, MouseEvent } from 'react';
 
-import { FeatureToggle } from '../../universal/config/feature-toggles';
-import { myThemasMenuItems } from '../config/thema';
-import { ThemaMenuItem } from '../config/thema-types';
-
 // Repeating conditions for accessible keyboard event
 // eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
 export function withKeyPress<T>(fn: Function, keyName: string = 'enter') {
@@ -51,33 +47,4 @@ export function getElementOnPageAsync(
     }
     checkIfElementIsInDOM(); // Initial check
   });
-}
-
-const REDACTED_CLASS = 'cobrowse-redacted';
-export const hasRedactedClass = (
-  themaId: string,
-  scope: Required<ThemaMenuItem>['redactedScope']
-) => {
-  if (!FeatureToggle.cobrowseIsActive) {
-    return false;
-  }
-  const redactedScope = myThemasMenuItems.find(
-    (item) => item.id === themaId
-  )?.redactedScope;
-  return (
-    redactedScope === 'full' ||
-    (redactedScope === 'content' && scope === 'content')
-  );
-};
-export function getRedactedClass(
-  themaId?: string | null,
-  scope: Required<ThemaMenuItem>['redactedScope'] = 'full'
-) {
-  if (
-    !FeatureToggle.cobrowseIsActive ||
-    (themaId && !hasRedactedClass(themaId, scope))
-  ) {
-    return '';
-  }
-  return REDACTED_CLASS;
 }
