+++ src/client/pages/Thema/HLI/HLIStadspasDetail.tsx
@@ -40,7 +40,7 @@ import {
 import { PageHeadingV2 } from '../../../components/PageHeading/PageHeadingV2';
 import { Spinner } from '../../../components/Spinner/Spinner';
 import { TableV2 } from '../../../components/Table/TableV2';
-import { getRedactedClass } from '../../../helpers/utils';
+import { getRedactedClass } from '../../../helpers/cobrowse';
 import { useDataApi } from '../../../hooks/api/useDataApi';
 import { useSmallScreen } from '../../../hooks/media.hook';
 import { useAppStateGetter } from '../../../hooks/useAppState';
