+++ src/client/pages/Thema/Profile/private/useContactmomenten.hook.tsx
@@ -17,7 +17,7 @@ import styles from './ProfilePrivate.module.scss';
 import { isLoading, isError } from '../../../../../universal/helpers/api';
 import { MaRouterLink } from '../../../../components/MaLink/MaLink';
 import { ThemaMenuItemTransformed } from '../../../../config/thema-types';
-import { getRedactedClass } from '../../../../helpers/utils';
+import { getRedactedClass } from '../../../../helpers/cobrowse';
 import { useAppStateGetter } from '../../../../hooks/useAppState';
 import {
   useThemaBreadcrumbs,
